+++ services/web-app/data/design-kits.js
@@ -30,16 +30,16 @@ const designKitAllowList = {
     maintainer: 'carbon'
   },
   'ibm-design-language-sketch': {
-    maintainer: 'community'
+    maintainer: 'ibm-brand'
   },
   'ibm-icons-16-20-sketch': {
-    maintainer: 'carbon'
+    maintainer: 'ibm-brand'
   },
   'ibm-icons-24-32-sketch': {
-    maintainer: 'carbon'
+    maintainer: 'ibm-brand'
   },
   'ibm-grid-sketch': {
-    maintainer: 'community'
+    maintainer: 'carbon'
   },
   'carbon-shell-sketch': {
     maintainer: 'carbon'
@@ -57,26 +57,22 @@ const designKitAllowList = {
     maintainer: 'carbon'
   },
   'carbon-white-adobe-xd': {
-    maintainer: 'community'
+    maintainer: 'carbon'
   },
   'carbon-g10-adobe-xd': {
-    maintainer: 'community'
+    maintainer: 'carbon'
   },
   'carbon-g90-adobe-xd': {
-    maintainer: 'community'
+    maintainer: 'carbon'
   },
   'carbon-g100-adobe-xd': {
-    maintainer: 'community'
+    maintainer: 'carbon'
   },
   'ibm-icons-adobe-xd': {
-    maintainer: 'community'
+    maintainer: 'ibm-brand'
   },
-  'carbon-g10-axure': {
-    maintainer: 'community'
-  },
-  'carbon-template-axure': {
-    maintainer: 'community'
-  },
+  'carbon-g10-axure': {},
+  'carbon-template-axure': {},
   'data-viz-sketch': {
     maintainer: 'carbon'
   },
@@ -91,10 +87,10 @@ const designKitAllowList = {
   'ibm-grid-mobile-sketch': {},
   'mobile-patterns-sketch': {},
   'ibm-accessibility-sketch': {
-    maintainer: 'IBM-accessibility'
+    maintainer: 'ibm-accessibility'
   },
   'ibm-accessibility-figma': {
-    maintainer: 'community'
+    maintainer: 'ibm-accessibility'
   },
   'carbon-mid-fi-sketch': {},
   'carbon-wireframe-invision-freehand': {
@@ -119,5 +115,14 @@ const designKitAllowList = {
     maintainer: 'ibm-cloud'
   }
 }
+
+/**
+ * Design kits with no maintaining team default to `community`.
+ */
+Object.keys(designKitAllowList).forEach((library) => {
+  if (!designKitAllowList[library].maintainer) {
+    designKitAllowList[library].maintainer = 'community'
+  }
+})
 
 export { designKitAllowList, designKitSources }